Add userExtractor middleware

+++ controller/blogs.js
@@ -2,6 +2,7 @@ const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
 const User = require("../models/user")
 const jwt = require("jsonwebtoken")
+const middleware = require("../utils/middleware")
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog
@@ -9,15 +10,15 @@ blogsRouter.get("/", async (request, response) => {
   response.json(blogs)
 })
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
   const body = request.body
   let decodedToken = undefined
   try {
     decodedToken = jwt.verify(request.token, process.env.SECRET)
-  } catch(e) {
+  } catch (e) {
     return response.status(401).json({ error: "token invalid" })
   }
-  const user = await User.findById(decodedToken.id)
+  const user = request.user
 
   const blog = new Blog({
     ...body,
@@ -31,8 +32,15 @@ blogsRouter.post("/", async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+blogsRouter.delete("/:id", middleware.userExtractor, async (request, response) => {
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog === null) return response.status(204).end()
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+  }
   response.status(204).end()
 })
 
@@ -45,8 +53,8 @@ blogsRouter.put("/:id", async (request, response) => {
     { new: true, runValidators: true, context: "query" }
   )
 
-  updatedBlog === null 
-    ? response.status(404).json({ error: "Blog does not exist"})
+  updatedBlog === null
+    ? response.status(404).json({ error: "Blog does not exist" })
     : response.json(updatedBlog)
 })
